[OAP-1727][POAE7-358] Spark integration: Memory Spill to PMem (#1725)
* [OAP-1727][POAE7-358] Spark integration: Memory Spill to PMem

* rever unnecessary change in BytesToBytesMap

* retrigger CI

+++ oap-spark/src/main/java/org/apache/spark/unsafe/map/BytesToBytesMap.java
@@ -25,6 +25,10 @@ import java.util.LinkedList;
 
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.io.Closeables;
+import com.intel.oap.common.storage.stream.ChunkInputStream;
+import com.intel.oap.common.storage.stream.DataStore;
+import org.apache.spark.internal.config.package$;
+import org.apache.spark.memory.PMemManagerInitializer;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -393,10 +397,23 @@ public final class BytesToBytesMap extends MemoryConsumer {
     }
 
     private void handleFailedDelete() {
-      // remove the spill file from disk
+      // remove the spill file from disk or pmem
+      boolean pMemSpillEnabled = SparkEnv.get() != null && (boolean) SparkEnv.get().conf().get(
+              package$.MODULE$.MEMORY_SPILL_PMEM_ENABLED());
       File file = spillWriters.removeFirst().getFile();
-      if (file != null && file.exists() && !file.delete()) {
-        logger.error("Was unable to delete spill file {}", file.getAbsolutePath());
+      if(pMemSpillEnabled == true) {
+        try {
+          ChunkInputStream cis = ChunkInputStream.getChunkInputStreamInstance(file.toString(),
+                  new DataStore(PMemManagerInitializer.getPMemManager(),
+                          PMemManagerInitializer.getProperties()));
+          cis.free();
+        } catch (IOException e) {
+          logger.debug(e.toString());
+        }
+      } else {
+        if (file != null && file.exists() && !file.delete()) {
+          logger.error("Was unable to delete spill file {}", file.getAbsolutePath());
+        }
       }
     }
   }
@@ -818,12 +835,42 @@ public final class BytesToBytesMap extends MemoryConsumer {
     }
     assert(dataPages.isEmpty());
 
+    deleteSpillFiles();
+  }
+
+  /**
+   * Deletes any spill files created by this consumer.
+   */
+  private void deleteSpillFiles() {
+    boolean pMemSpillEnabled = SparkEnv.get() != null && (boolean) SparkEnv.get().conf().get(
+            package$.MODULE$.MEMORY_SPILL_PMEM_ENABLED());
+    if (pMemSpillEnabled == true)
+      deletePMemSpillFiles();
+    else
+      deleteDiskSpillFiles();
+  }
+
+  private void deleteDiskSpillFiles() {
     while (!spillWriters.isEmpty()) {
       File file = spillWriters.removeFirst().getFile();
       if (file != null && file.exists()) {
         if (!file.delete()) {
           logger.error("Was unable to delete spill file {}", file.getAbsolutePath());
         }
+      }
+    }
+  }
+
+  private void deletePMemSpillFiles() {
+    while (!spillWriters.isEmpty()) {
+      File file = spillWriters.removeFirst().getFile();
+      try {
+        ChunkInputStream cis = ChunkInputStream.getChunkInputStreamInstance(file.toString(),
+                new DataStore(PMemManagerInitializer.getPMemManager(),
+                        PMemManagerInitializer.getProperties()));
+        cis.free();
+      } catch (IOException e) {
+        logger.debug(e.toString());
       }
     }
   }
